Update app-install.sh
On some systems the original script will not execute and ask for the package Python3-flask even if the package is already installed. This update fix this problem and makes it easy to install adsb image on other linux systems. (Testet on Repka Pi3 with Ubuntu 20.04.06)

RTL-SDR may not work properly if Docker-Install is not installed:
https://github.com/sdr-enthusiasts/docker-install -> only this!